// packages/shared/src/generated/inputTypeSchemas/DashboardSnapshotSumOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const DashboardSnapshotSumOrderByAggregateInputSchema: z.ZodType<Prisma.DashboardSnapshotSumOrderByAggregateInput> = z.strictObject({
  totalEvents: z.lazy(() => SortOrderSchema).optional(),
  totalUsers: z.lazy(() => SortOrderSchema).optional(),
  publishedEvents: z.lazy(() => SortOrderSchema).optional(),
  draftEvents: z.lazy(() => SortOrderSchema).optional(),
  cancelledEvents: z.lazy(() => SortOrderSchema).optional(),
  totalCapacity: z.lazy(() => SortOrderSchema).optional(),
  usedCapacity: z.lazy(() => SortOrderSchema).optional(),
});

export default DashboardSnapshotSumOrderByAggregateInputSchema;